setFormula implementation
